ManufacturingAgents/requirements.txt
google-adk[database]==1.6.1
pandas
